+++ src/main/java/com/almondtools/rexlex/automaton/BaseListener.java
@@ -1,17 +1,17 @@
 package com.almondtools.rexlex.automaton;
 
 import com.almondtools.rexlex.TokenType;
-import com.almondtools.rexlex.io.CharProvider;
+import com.almondtools.stringsandchars.io.CharProvider;
 
 public class BaseListener implements AutomatonMatcherListener {
 
 	@Override
-	public boolean reportMatch(CharProvider chars, int start, TokenType accepted) {
+	public boolean reportMatch(CharProvider chars, long start, TokenType accepted) {
 		return false;
 	}
 
 	@Override
-	public boolean recoverMismatch(CharProvider chars, int start) {
+	public boolean recoverMismatch(CharProvider chars, long start) {
 		chars.move(start + 1);
 		return false;
 	}
