+++ src/main/java/com/almondtools/rexlex/automaton/MatchCollector.java
@@ -4,8 +4,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 import com.almondtools.rexlex.TokenType;
-import com.almondtools.rexlex.io.CharProvider;
 import com.almondtools.rexlex.pattern.Match;
+import com.almondtools.stringsandchars.io.CharProvider;
 
 public class MatchCollector implements AutomatonMatcherListener {
 
@@ -16,8 +16,8 @@ public class MatchCollector implements AutomatonMatcherListener {
 	}
 	
 	@Override
-	public boolean reportMatch(CharProvider chars, int start, TokenType accepted) {
-		int end = chars.current();
+	public boolean reportMatch(CharProvider chars, long start, TokenType accepted) {
+		long end = chars.current();
 		String text = chars.slice(start, end);
 		if (chars.finished()) {
 			matches.add(new Match(start, text, accepted));
@@ -26,7 +26,7 @@ public class MatchCollector implements AutomatonMatcherListener {
 	}
 
 	@Override
-	public boolean recoverMismatch(CharProvider chars, int start) {
+	public boolean recoverMismatch(CharProvider chars, long start) {
 		return true;
 	}
